# Conflicts:
#	src/main/java/de/uniks/stpmon/team_m/controller/EncounterController.java

+++ src/main/java/de/uniks/stpmon/team_m/dto/Layer.java
@@ -4,7 +4,7 @@ import java.util.List;
 
 public record Layer(
         List<Chunk> chunks,
-        List<Integer> data,
+        List<Long> data,
         int height,
         int id,
         String name,
